refactor: move the data folder from app to src
