+++ carrerait-roadmap/src/data/aspirantQuestions.ts
