+++ carrerait-roadmap/src/data/itWorkerConditionalResponses.ts
