+++ carrerait-roadmap/src/data/itWorkerQuestions.ts
