+++ carrerait-roadmap/src/data/resources.ts
